\name{laser}
\alias{laser}
\title{Bellcore Experimental Laser Data}
\usage{data(laser)}
\description{
    This data came from an investigation of an experimental laser at
    Bellcore.  It was a tunable laser, in the sense that both its
    wavelength and power output were controllable.}
\details{
    Rotation helped the experimental physicists to characterize the
    laser, which turned out not to be a very good one, due to its
    unstable operating region.

    This data initially came to the statistics research group when
    Janette Cooper asked Paul Tukey to help her analyze the data she had
    collected to describe the laser.}
\format{
    A data frame with observations on the variables \code{If} and
    \code{Ir} (the currents applied to the front and rear of the laser),
    and \code{power} and \code{lambda} (the output power and
    wavelength).
}
\keyword{datasets}
